+++ assets/js/pages/production/show.js
@@ -2,7 +2,7 @@ import AJAX, {DELETE, GET} from "@app/ajax";
 import Form from "@app/form";
 import Modal from "@app/modal";
 import Camera from "@app/camera"
-import {displayAttachmentRequired} from './form'
+import {displayAttachmentRequired, openModalUpdateProductionRequestStatus} from '@app/pages/production/form'
 
 global.deleteProductionRequest = deleteProductionRequest;
 global.openModalEditProductionRequest = openModalEditProductionRequest;
@@ -30,6 +30,15 @@ $(function () {
 
     getStatusHistory(productionRequestId);
     getOperationHistory(productionRequestId);
+
+    const $modalUpdateProductionRequestStatus = $(`#modalUpdateProductionRequestStatus`);
+
+    $(document).on('click', '.open-modal-update-production-request-status', $modalUpdateProductionRequestStatus, (event) => {
+        openModalUpdateProductionRequestStatus($(this), $modalUpdateProductionRequestStatus, productionRequestId, () => {
+            window.location.reload();
+        })
+    });
+
 });
 
 function getStatusHistory(productionRequestId) {
