+++ src/Controller/ProductionRequest/ProductionRequestController.php
@@ -390,4 +390,50 @@ class ProductionRequestController extends AbstractController
             throw new BadRequestHttpException();
         }
     }
+
+    #[Route("/{productionRequest}/update-status-content", name: "update_status_content", options: ["expose" => true], methods: [self::GET])]
+    public function productionRequestUpdateStatusContent(EntityManagerInterface $entityManager,
+                                                         ProductionRequest      $productionRequest): JsonResponse {
+        $fixedFieldRepository = $entityManager->getRepository(FixedFieldStandard::class);
+
+        $html = $this->renderView('production_request/planning/update-status-form.html.twig', [
+            "productionRequest" => $productionRequest,
+            "fieldsParam" => $fixedFieldRepository->getByEntity(FixedFieldStandard::ENTITY_CODE_PRODUCTION),
+        ]);
+
+        return $this->json([
+            "success" => true,
+            "html" => $html
+        ]);
+    }
+
+    #[Route("/{productionRequest}/update-status", name: "update_status", options: ["expose" => true], methods: self::POST)]
+    public function updateStatus(EntityManagerInterface   $entityManager,
+                                 ProductionRequest        $productionRequest,
+                                 Request                  $request,
+                                 ProductionRequestService $productionRequestService): JsonResponse {
+
+        $productionRequestService->checkRoleForEdition($productionRequest);
+
+        $currentUser = $this->getUser();
+
+        $inputBag = new InputBag([
+            FixedFieldEnum::status->name => $request->request->get(FixedFieldEnum::status->name),
+            FixedFieldEnum::comment->name => $request->request->get(FixedFieldEnum::comment->name),
+        ]);
+
+        $oldStatus = $productionRequest->getStatus();
+        $productionRequestService->updateProductionRequest($entityManager, $productionRequest, $currentUser, $inputBag, $request->files);
+
+        $entityManager->flush();
+
+        if($oldStatus->getId() !== $productionRequest->getStatus()->getId()) {
+            $productionRequestService->sendUpdateStatusEmail($productionRequest);
+        }
+
+        return $this->json([
+            "success" => true,
+            "msg" => "La demande de production a été modifiée avec succès.",
+        ]);
+    }
 }
