+++ templates/production_request/modal/updateStatus.html.twig
@@ -1,0 +1,12 @@
+{% include 'utils/modal/empty-modal.html.twig' with {
+    id: 'modalUpdateProductionRequestStatus',
+    header: {
+        title: 'Changer de statut'
+    },
+    footer: {
+        elements: [
+            {classes: 'btn btn-outline-secondary', dismiss: true, text: trans('Général', null, 'Modale', 'Annuler')},
+            {classes: 'btn btn-success', text: trans('Général', null, 'Modale', 'Enregistrer'), type: 'submit'},
+        ],
+    }
+} %}
