+++ templates/production_request/planning/card.html.twig
@@ -23,10 +23,7 @@
          style="background: {{ color }}50; border-left-color: {{ color }};">
         <div class="row">
             <div class="{{ emergency ? 'col-8' : 'col-12' }}">
-                <div class="rounded px-2 py-1 d-inline-block bg-white {{ canUpdateStatus ? "prevent-default" }}"
-                     {% if canUpdateStatus %}
-                        onclick="openModalUpdateProductionRequestStatus($(this))"
-                     {% endif %}>
+                <div class="rounded px-2 py-1 d-inline-block bg-white {{ canUpdateStatus ? "prevent-default" }} {{ canUpdateStatus  ? "open-modal-update-production-request-status"}}">
                     {{ productionRequest.status|format_helper("status") }}
                 </div>
                 <div class="mt-1">
